+++ prisma/migrations/20260529173037_add_program_federation_membership_restriction/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Program" ADD COLUMN     "hasFederationMembershipRestriction" BOOLEAN NOT NULL DEFAULT false;